Separate bib_item into a concern
